+++ pages/index.tsx
@@ -2,14 +2,10 @@ import { useRouter } from 'next/router';
 import React from 'react';
 import Layout from '../src/features/common/Layout';
 import ProjectsList from '../src/features/public/Donations/screens/Projects';
-import Head from 'next/head';
-import tenantConfig from '../tenant.config';
-import getImageUrl from '../src/utils/getImageURL';
 import { getAllProjects } from '../src/utils/apiRequests/getAllProjects';
 import { getSingleProject } from '../src/utils/apiRequests/getSingleProject';
 import GetProjectMeta from '../src/utils/getMetaTags/GetProjectMeta';
-
-const config = tenantConfig();
+import GetAllProjectsMeta from '../src/utils/getMetaTags/GetAllProjectsMeta';
 
 interface Props {
   initialized: Boolean;
@@ -74,27 +70,7 @@ export default function Donate(initialized: Props) {
       {project ? (
           <GetProjectMeta project={project} />
       ) : (
-        <Head>
-          <title>{config.meta.title}</title>
-          <meta property="og:site_name" content={config.meta.title} />
-          <meta
-            property="og:url"
-            content={`${process.env.SCHEME}://${config.tenantURL}`}
-          />
-          <meta property="og:title" content={config.meta.title} />
-          <meta property="og:description" content={config.meta.description} />
-          <meta name="description" content={config.meta.description} />
-          <meta property="og:type" content="website" />
-          <meta property="og:image" content={config.meta.image} />
-          {config.tenantName === 'planet' ? (
-            <link rel="alternate" href="android-app://org.pftp/projects" />
-          ) : null}
-          <meta name="twitter:card" content="summary" />
-          <meta name="twitter:title" content={config.meta.title} />
-          <meta name="twitter:site" content={config.meta.twitterHandle} />
-          <meta name="twitter:url" content={config.tenantURL} />
-          <meta name="twitter:description" content={config.meta.description} />
-        </Head>
+        <GetAllProjectsMeta/>
       )}
       {initialized ? (
         <Layout>
